// src/interfaces/class_data.tsx
export interface ClassData {
  id: number;
  dc_nombre_clase: string;
  dc_horario: string;
  df_fecha: string;
  nb_cupos_disponibles: any;
  dc_descripcion: string;
  dc_imagen_url: string;
  tb_gimnasio_id: any;
  id_categoria: any;
  tb_clase_estado_id: any;
  tb_arte_marcial_id: any;
  tb_profesor_id: any;
  hora_inicio: any;
  hora_fin: any;
}